+++ src/AbstractCommonList.php
@@ -69,7 +69,7 @@ abstract class AbstractCommonList implements \Iterator, \Countable
      * @return AbstractNode|null
      * 
      */
-    public function getHeadNode(int $offset = 0): ?AbstractNode
+    public function head(int $offset = 0): ?AbstractNode
     {
         if ($offset === 0)
             return $this->head;